requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-ses==3.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.5.0
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
sendgrid==6.10.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
urllib3==1.26.15
